refactor: move indexes related types to src/lib
