+++ src/views/default-page-view/default-page-view.astro
@@ -2,10 +2,10 @@
 import type { ComponentProps } from "astro/types";
 import CollectionMeta from "../../components/organisms/collection-meta/collection-meta.astro";
 import PageLayout from "../../components/templates/page-layout/page-layout.astro";
+import type { RoutableCollectionKey } from "../../lib/astro/collections/types";
 import { getWebPageGraph } from "../../lib/schema-dts/graphs/webpage-graph";
 import { getBreadcrumb } from "../../services/breadcrumb";
 import type { QueriedEntry } from "../../services/collections";
-import type { RoutableCollectionKey } from "../../types/routing";
 
 type Props = Partial<ComponentProps<typeof PageLayout>> & {
   entry: QueriedEntry<RoutableCollectionKey>;
